requirements.txt
lxml==4.6.3
numpy==1.20.3
Pillow==8.2.0
pydantic==1.8.2
typing-extensions==3.10.0.0
